+++ lib/recording_studio_oauth/oauth_client_form.rb
@@ -10,7 +10,10 @@ module RecordingStudioOauth
         :use_central_relay,
         :allowed_return_patterns,
         :exact_return_urls,
-        :allow_registration
+        :allow_registration,
+        :session_token_provider,
+        :session_token_audience,
+        :session_token_secret
       )
     end
 
@@ -42,7 +45,10 @@ module RecordingStudioOauth
         redirect_uris: lines(params[:redirect_uris]),
         use_central_relay: relay_flag,
         allowed_return_patterns: lines(params[:allowed_return_patterns]),
-        exact_return_urls: lines(params[:exact_return_urls])
+        exact_return_urls: lines(params[:exact_return_urls]),
+        session_token_provider: params[:session_token_provider],
+        session_token_audience: params[:session_token_audience],
+        session_token_secret: params[:session_token_secret]
       }
     end